Extrace data exporter as a submodule
