+++ docs/source/api_reference/data_exporters/base.rst
@@ -1,0 +1,9 @@
+DataExporter
+=======================
+
+.. autoclass:: sdgx.data_exporters.base.DataExporter
+   :members:
+   :undoc-members:
+   :inherited-members:
+   :show-inheritance:
+   :private-members:
