+++ docs/source/api_reference/data_exporters/csv_exporter.rst
@@ -1,0 +1,10 @@
+CsvExporter
+=====================================
+
+
+.. autoclass:: sdgx.data_exporters.csv_exporter.CsvExporter
+    :members:
+    :undoc-members:
+    :inherited-members:
+    :show-inheritance:
+    :private-members:
