+++ docs/source/api_reference/data_exporters/extension.rst
@@ -1,0 +1,11 @@
+.. _api_reference/data-exporters-extension:
+
+Extension hookspec
+============================
+
+.. automodule:: sdgx.data_exporters.extension
+   :members:
+   :undoc-members:
+   :inherited-members:
+   :show-inheritance:
+   :private-members:
