+++ docs/source/api_reference/data_exporters/manager.rst
@@ -1,0 +1,9 @@
+DataExporterManager
+=================================
+
+.. autoclass:: sdgx.data_exporters.manager.DataExporterManager
+   :members:
+   :undoc-members:
+   :inherited-members:
+   :show-inheritance:
+   :private-members:
