+++ docs/source/api_reference/index.rst
@@ -14,6 +14,7 @@ API Reference
     Data Processor <data_processors/index>
     Models <models/index>
     Metadata and Inspectors <data_models/index>
+    DataExporter for exporting data <exporters/index>
     Manager <manager>
     Exceptions <exceptions>
     Utils <utils>
