+++ example/extension/dummyexporter/pyproject.toml
@@ -23,5 +23,5 @@ ignore = [".*"]
 [tool.hatch.version]
 path = "dummyexporter/__init__.py"
 
-[project.entry-points."sdgx.cli.exporter"]
+[project.entry-points."sdgx.cli.DataExporter"]
 dummyexporter = "dummyexporter.dummyexporter"
